+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -8,6 +8,12 @@ CMake Deprecation Warning:
   use the cmake-file-api(7) to view CMake-generated project build trees.
 
 
--- Configuring done (1.2s)
--- Generating done (0.1s)
+-- The C compiler identification is GNU 13.2.0
+-- Detecting C compiler ABI info
+-- Detecting C compiler ABI info - done
+-- Check for working C compiler: /usr/bin/cc - skipped
+-- Detecting C compile features
+-- Detecting C compile features - done
+-- Configuring done (0.4s)
+-- Generating done (0.0s)
 -- Build files have been written to: /home/shuta/IdeaProjects/MMDMotionMaker/cmake-build-debug
