+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Jan 24 23:31 JST
+Start testing: Jan 27 14:36 JST
 ----------------------------------------------------------
-End testing: Jan 24 23:31 JST
+End testing: Jan 27 14:36 JST
